Improvements
Now the abort and bus errors are captured to prevent problems in the tester, a message is displayed if there is one. Prevented the execution of the program with fsanitize since it causes problems in certain versions of gcc by preloading the library before asan. It is also checked that the program is compiled with debugging enabled -g, since without this the error/leak traces cannot be shown.